Move fast scrolling TLC to com.google.android.horologist.compose.layout.m3
